requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.0
numpy==1.24.0
matplotlib==3.7.0
yfinance==0.2.0
scikit-learn==1.3.0
tensorflow==2.15.0
seaborn==0.13.0
plotly==5.18.0
ta==0.10.0
python-binance==1.0.19
ccxt==4.1.13
colorama==0.4.6
tqdm==4.66.1
rich==13.7.0 